requirements.txt
jinja2
huggingface_hub
transformers
python-dotenv
